Fluentd changed to a multi-processes service, and suricata is integrated into a single container,delete grafana and shutdown flowscan,AI,packetbeat....
