+++ SecBuzzerESM/Crontab/docker-compose.yml
@@ -8,10 +8,10 @@ services:
       context: "./docker_image/"
       dockerfile: "Dockerfile"
     volumes:
-      #- "$FLUENTD_LOG_PATH:/fluentd_log"
+      - "$FLUENTD_LOG_PATH:/fluentd_log"
       - "$ES_VOLUME_PATH:/es_volume"
       - "$SURICATA_LOG_PATH:/suricata_log"
-      - "../Suricata/suricata/rules:/Suricata_rules/rules"
+      - "../Suricata/suricata/rules/:/Suricata_rules"
       - "/usr/share/zoneinfo/ROC:/etc/localtime:ro"
       - "./scripts:/scripts:ro"
       - "/etc/hostname:/hostname:ro"
@@ -19,6 +19,8 @@ services:
       - ../SecBuzzerESM.env:/.env:ro
       - "$FLUENTD_BUFFER_PATH/Logs/:/Logs/"
       - "$FLUENTD_LOG_PATH:/LogsCompress"
+    environment:
+      DEV: ${DEV_MODE:-no}
     networks:
       - esm
     restart: unless-stopped
