+++ SecBuzzerESM/Crontab/scripts/Suricata_updater.sh
@@ -2,12 +2,12 @@
 
 sleep $[ ( $RANDOM % 15 ) + 1 ]s
 ESM_API=`cat /.env | grep API_KEY_VALUE | cut -d = -f 2`
-current_rules_version=`curl -X POST "https://test.api.secbuzzer.co/esmapi/web/file/fileVersion" -d "{'TypeCode': 'it'}" -H "Content-Type: application/json" -H "accept: */*" -H "authorization: $ESM_API" | cut -d : -f 2 | cut -d \" -f 2`
+current_rules_version=`curl -X POST "https://api.esm.secbuzzer.co/esmapi/web/file/fileVersion" -d "{'TypeCode': 'it'}" -H "Content-Type: application/json" -H "accept: */*" -H "authorization: $ESM_API" | cut -d : -f 2 | cut -d \" -f 2`
 local_rules_version=$(</tmp/local_rules_version)
 if [ "$current_rules_version" != "$local_rules_version" ]
 then
     echo "[*] New version found! rules update to $current_rules_version"
-    curl -o /tmp/rules.tgz "https://test.api.secbuzzer.co/esmapi/web/file/download/it/$current_rules_version" -H "accept: */*" -H "authorization: $ESM_API"
+    curl -o /tmp/rules.tgz "https://api.esm.secbuzzer.co/esmapi/web/file/download/it/$current_rules_version" -H "accept: */*" -H "authorization: $ESM_API"
     mkdir -p /tmp/rules
     tar zxf /tmp/rules.tgz -C /tmp/rules
     chown 1000:1000 /tmp/* -R
