+++ SecBuzzerESM/Crontab/scripts/daily.sh
@@ -15,21 +15,24 @@ curl -X DELETE "elasticsearch:9200/winlog-$winlog_remove_date"
 suricata_remove_date="$(date -d '-6 month' +%Y-%m)"
 curl -X DELETE "elasticsearch:9200/suricata-$suricata_remove_date"
 
-cic_remove_date="$(date -d '-7 days' +%Y%m%d)"
-curl -X DELETE "elasticsearch:9200/cic-$cic_remove_date"
+dtm_remove_date="$(date -d '-6 month' +%Y-%m)"
+curl -X DELETE "elasticsearch:9200/dtm-$dtm_remove_date"
+
+#cic_remove_date="$(date -d '-7 days' +%Y%m%d)"
+#curl -X DELETE "elasticsearch:9200/cic-$cic_remove_date"
 
 winlogbeat_remove_date="$(date -d '-7 days' +%Y%m%d)"
 curl -X DELETE "elasticsearch:9200/winlogbeat-$winlogbeat_remove_date"
 
-packetbeat_remove_date="$(date -d '-7 days' +%Y.%m.%d)"
-curl -X DELETE "elasticsearch:9200/packetbeat-$packetbeat_remove_date"
+#packetbeat_remove_date="$(date -d '-7 days' +%Y.%m.%d)"
+#curl -X DELETE "elasticsearch:9200/packetbeat-$packetbeat_remove_date"
 
-flowscan_remove_date="$(date -d '-6 month' +%Y-%m)"
-curl -X DELETE "elasticsearch:9200/flowscan-$flowscan_remove_date"
+#flowscan_remove_date="$(date -d '-6 month' +%Y-%m)"
+#curl -X DELETE "elasticsearch:9200/flowscan-$flowscan_remove_date"
 
-nmapscan_remove_date="$(date -d '-6 month' +%Y-%m)"
-curl -X DELETE "elasticsearch:9200/nmap-$nmapscan_remove_date"
-echo "[*] Daily check and remove index"
+#nmapscan_remove_date="$(date -d '-6 month' +%Y-%m)"
+#curl -X DELETE "elasticsearch:9200/nmap-$nmapscan_remove_date"
+#echo "[*] Daily check and remove index"
 
 lm_remove_date="$(date -d '-184 day' +%Y-%m-%d)"
 curl -X DELETE "elasticsearch:9200/lm-$lm_remove_date"
