+++ SecBuzzerESM/Fluentd/conf/fluent.conf.bak
@@ -1,0 +1,333 @@
+<system>
+  workers 6
+  root_dir /fluentd/Buffers/
+</system>
+
+<label @FLUENT_LOG>
+  <match **>
+    @type stdout
+  </match>
+</label>
+
+# suricata's worker
+<worker 0>
+  @include /fluentd/etc/suricata.conf
+</worker>
+
+<worker 1-2>
+  <source>
+    @type udp
+    tag syslog0
+    <parse>
+      @type none
+    </parse>
+    source_address_key _ip
+    source_hostname_key _hostname
+    port 514
+    bind 0.0.0.0
+    message_length_limit 1MB
+  </source>
+</worker>
+
+<worker 3-4>
+  <source>
+    @type udp
+    tag cef0
+    <parse>
+      @type none
+    </parse>
+    source_address_key _ip
+    source_hostname_key _hostname
+    port 515
+    bind 0.0.0.0
+    message_length_limit 1MB
+  </source>
+</worker>
+
+<worker 5>
+  <source>
+    @type forward
+    port 24224
+    bind 0.0.0.0
+    resolve_hostname true
+    source_address_key _ip
+    source_hostname_key _hostname
+  </source>
+</worker>
+
+<match syslog0>
+@type copy
+  <store>
+    @type file
+    path /fluentd/Buffers/Logs/syslog/${_ip}/%Y-%m-%d
+    append true
+    <format>
+        @type json
+    </format>
+    <inject>
+      time_key @timestamp
+      time_type string
+      time_format %Y-%m-%dT%H:%M:%S.%N%z
+    </inject>
+    <buffer time, _ip>
+      @type file
+      path /fluentd/Buffers/syslog_Buffers/
+      timekey 1d
+      flush_interval 1m
+      flush_at_shutdown true
+      flush_mode interval
+      flush_thread_count 4
+      timekey_use_utc false 
+      chunk_limit_size 2M
+      total_limit_size 5G
+    </buffer>
+  </store>
+  <store>
+    @type rewrite_tag_filter
+    <rule>
+    key message
+    pattern \w+
+    tag syslog1
+    </rule>
+  </store>
+</match>
+
+<match cef0>
+@type copy
+  <store>
+    @type file
+    path /fluentd/Buffers/Logs/cef/${_ip}/%Y-%m-%d
+    append true
+    <format>
+        @type json
+    </format>
+    <inject>
+      time_key @timestamp
+      time_type string
+      time_format %Y-%m-%dT%H:%M:%S.%N%z
+    </inject>
+    <buffer time, _ip>
+      @type file
+      path /fluentd/Buffers/cef_Buffers/
+      timekey 1d
+      flush_interval 1m
+      flush_at_shutdown true
+      flush_mode interval
+      flush_thread_count 4
+      timekey_use_utc false 
+      chunk_limit_size 2M
+      total_limit_size 5G
+    </buffer>
+  </store>
+  <store>
+    @type rewrite_tag_filter
+    <rule>
+    key message
+    pattern \w+
+    tag cef1
+    </rule>
+  </store>
+</match>
+
+<filter syslog1.**>
+  @type parser
+  key_name message
+  emit_invalid_record_to_error false
+  reserve_data true
+  <parse>
+    @type regexp
+    expression /^(?:<(?<pri>[\d]+)>)?(?<timestamp>[\w]+(?: [\d]{2}|  [\d]{1}) [\d]{2}:[\d]{2}:[\d]{2}) (?:Message forwarded from )?(?<hostname>[^ :]*) ?(?: |: )(?<ident>[^ :\[]*)(?:\[(?<pid>[0-9]+)\])?(?:[^\:]*\:)? *(?<message>.*)$/
+  </parse>
+</filter>
+
+<filter syslog1.**>
+  @type parser
+  key_name message
+  emit_invalid_record_to_error false
+  reserve_data true
+  <parse>
+    @type regexp
+    expression /\A\<(?<pri>[0-9]{1,3})\>[1-9]\d{0,2} (?<timestamp>[\d]{4}-[\d]{2}-[\d]{2}T[\d]{2}:[\d]{2}:[\d]{2}(?:[\d]{0}|.[\d]{6})\+[\d]{2}:[\d]{2}) (?<hostname>[!-~]{1,255}) (?<ident>[!-~]{1,48}) (?<pid>[!-~]{1,128}) (?<msgid>[!-~]{1,32}) (?<extradata>(?:\-|(?:\[.*?(?<!\\)\])+))(?: (?<message>[\w\W\n]+))?\z/
+  </parse>
+</filter>
+
+# pri @timestamp hostname Parser
+<filter cef1.**>
+  @type parser
+  key_name message
+  emit_invalid_record_to_error false
+  reserve_data true
+  <parse>
+    @type regexp
+    expression /^(?:(?:<(?<pri>[\d]+)>)?(?<timestamp>(?:\w{3}\s+\d+\s+\d+:\d+:\d+)|(?:d{4}\-\d{2}\-\d{2}T\d{2}\:\d{2}\:\d{2}\+\d{2}\:\d{2})|(?:\w{3}\s+\d+\s\d+\s\d+:\d+:\d+)) (?<hostname>[^\s\|]+))/
+  </parse>
+</filter>
+
+# CEF head Parser
+<filter cef1.**>
+  @type parser
+  key_name message
+  reserve_data true
+  <parse>
+    @type regexp
+    expression /CEF:(?:\s)?(?<cef_version>[^\|]+)\|(?<cef_device_vendor>[^\|]+)\|(?<cef_device_product>[^\|]+)\|(?<cef_device_version>[^\|]+)\|(?<cef_device_event_class_id>[^\|]+)\|(?<cef_name>[^\|]+)\|(?<cef_severity>[^\|]+)\|/
+  </parse>
+</filter>
+
+#kvp Parser
+<filter cef1.**>
+    @type kvp
+    parse_key message
+    remove_parse_key true
+    pattern (\w+)=(.*?)(?= [\w]+=|$)
+</filter>
+
+#<match cef1.**>
+#  @type rewrite_tag_filter
+#  <rule>
+#    key cef_device_vendor
+#    pattern (Palo Alto)
+#    tag cef2
+#  </rule>
+#  <rule>
+#    key cef_device_vendor
+#    pattern \w+
+#    tag cef3
+#  </rule>
+#</match>
+
+#<match cef2.**>
+#  @type rewrite_tag_filter
+#  <rule>
+#    key cef_name
+#    pattern (^THREAT$|^threat$)
+#    tag cef3
+#  </rule>
+#  <rule>
+#    key cef_name
+#    pattern \w+
+#    tag clear
+#  </rule>
+#</match>
+
+#<match clear.**>
+#  @type null
+#</match>
+
+#log_time Parser
+<filter cef1.** syslog1.**>
+  @type record_transformer
+  auto_typecast true
+  enable_ruby true
+  <record>
+    pri_parser ${if record["pri"].nil?;"remove";else;"parser";end;}
+    log_time ${if record["timestamp"].length == 15 then Time.strptime(record["timestamp"], "%b %d %H:%M:%S").strftime('%Y-%m-%dT%H:%M:%S.%N%:z') elsif record["timestamp"].length == 20 then (Time.parse(record["timestamp"]) + 28800).strftime('%Y-%m-%dT%H:%M:%S.%N%:z') elsif record["timestamp"].length == 32 then Time.strptime(record["timestamp"], "%Y-%m-%dT%H:%M:%S.%N%:z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z') elsif record["timestamp"].length == 25 then Time.strptime(record["timestamp"], "%Y-%m-%dT%H:%M:%S%:z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z') end}
+  </record>
+</filter>
+
+<match cef3.**>
+  @type rewrite_tag_filter
+  <rule>
+    key pri_parser
+    pattern (remove)
+    tag cef_no_pri
+  </rule>
+  <rule>
+    key pri_parser
+    pattern (parser)
+    tag cef_parser_pri
+  </rule>
+</match>
+
+<match syslog1.**>
+  @type rewrite_tag_filter
+  <rule>
+    key pri_parser
+    pattern (remove)
+    tag syslog_no_pri
+  </rule>
+  <rule>
+    key pri_parser
+    pattern (parser)
+    tag syslog_parser_pri
+  </rule>
+</match>
+
+#pri Parser
+<filter cef_parser_pri.** syslog_parser_pri.**>
+  @type record_transformer
+  auto_typecast true
+  enable_ruby true
+  <record>
+    facility ${(record["pri"].to_i / 8).to_s}
+    severity ${(record["pri"].to_i % 8).to_s}
+  </record>
+</filter>
+
+<match cef_parser_pri.** cef_no_pri.**>
+  @type rewrite_tag_filter
+  <rule>
+    key pri_parser
+    pattern \w+
+    tag cef
+  </rule>
+</match>
+
+<match syslog_parser_pri.** syslog_no_pri.**>
+  @type rewrite_tag_filter
+  <rule>
+    key pri_parser
+    pattern \w+
+    tag syslog
+  </rule>
+</match>
+
+<filter **>
+  @type record_transformer
+  <record>
+    _log_type ${tag}
+  </record>
+  remove_keys pri_parser
+</filter>
+
+<filter eventlog.**>
+  @type record_transformer
+  auto_typecast true
+  enable_ruby true
+  <record>
+    log_time ${Time.strptime(record["time_generated"], "%Y-%m-%d %H:%M:%S %z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z')}
+  </record>
+</filter>
+
+<match **>
+@type copy
+  # <store>
+  #   @type stdout
+  # </store>
+  <store>
+    @type elasticsearch
+    host localhost
+    port 19200
+    index_name lm-%Y-%m-%d
+    include_timestamp true
+    reload_on_failure true
+    reconnect_on_error true
+    with_transporter_log true
+    reload_connections false
+    request_timeout 15s 
+    pipeline my_timestamp_pipeline
+    suppress_type_name true
+    <buffer tag, time>
+      @type file # or memory
+      path /fluentd/Buffers/ES_Buffers/
+      flush_mode interval
+      flush_interval 30s
+      flush_thread_count 4
+      timekey 1d
+      chunk_limit_size 15M
+      total_limit_size 5G # default is 64G
+      timekey_use_utc false
+      retry_max_interval 30s
+    </buffer>
+  </store>
+</match>
