+++ SecBuzzerESM/Fluentd/conf/in_fluentd.conf
@@ -1,0 +1,28 @@
+<source>
+  @type tail
+  path /fluentd/suricata/eve*.json
+  pos_file /fluentd/Buffers/suricata.pos
+  read_from_head true
+  tag suricata
+  <parse>
+    @type json
+  </parse>
+</source>
+
+<match suricata>
+  @type forward
+  num_threads 6
+  buffer_type file
+  buffer_queue_limit 1024
+  buffer_chunk_limit 2m
+  flush_interval 2s
+  buffer_path /fluentd/buffer/input.buffer
+  <server>
+    host 127.0.0.1
+    port 24000
+  </server>
+  <server>
+    host 127.0.0.1
+    port 24001
+  </server>
+</match>
