+++ SecBuzzerESM/Fluentd/conf/multiprocess.conf
@@ -1,0 +1,18 @@
+<source>
+  @type multiprocess
+  <process>
+    cmdline -c /fluentd/etc/in_fluentd.conf --log /var/log/fluentd/in_fluentd.conf.log
+    sleep_before_start 1s
+    sleep_before_shutdown 5s
+  </process>
+  <process>
+    cmdline -c /fluentd/etc/out_it_fluentd.conf --log /var/log/fluentd/out_it_fluentd.conf.log
+    sleep_before_start 1s
+    sleep_before_shutdown 5s
+  </process>
+  <process>
+    cmdline -c /fluentd/etc/out_dtm_fluentd.conf --log /var/log/fluentd/out_dtm_fluentd.conf.log
+    sleep_before_start 1s
+    sleep_before_shutdown 5s
+  </process>
+</source>
