+++ SecBuzzerESM/Fluentd/conf/out_it_fluentd.conf
@@ -1,0 +1,53 @@
+<source>
+  @type forward
+  port 24000
+  bind 0.0.0.0
+  tag suricata
+</source>
+
+<filter suricata>
+  @type grep
+  <exclude>
+    key $.alert.signature
+    pattern /DTM /
+  </exclude>
+</filter>
+
+<filter suricata>
+  @type record_transformer
+  auto_typecast true
+  enable_ruby true
+  <record>
+    log_time ${Time.strptime(record["timestamp"], "%Y-%m-%dT%H:%M:%S.%N%z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z')}
+    _log_type traffic
+    module suricata
+    dump_status 0
+    ticket 0
+  </record>
+</filter>
+
+<match suricata>
+  @type elasticsearch
+  host localhost
+  port 19200
+  index_name suricata-%Y-%m
+  include_timestamp true
+  reload_on_failure true
+  reconnect_on_error true
+  with_transporter_log true
+  reload_connections false
+  request_timeout 15s
+  pipeline my_timestamp_pipeline
+  suppress_type_name true
+  <buffer time>
+    @type file
+    timekey 1d
+    flush_mode interval
+    flush_interval 30s
+    retry_max_interval 30s
+    path /fluentd/Buffers/ES_Buffers/it_output.buffer
+    chunk_limit_size 15M
+    total_limit_size 5G # default is 64G
+    timekey_use_utc false
+  </buffer>
+</match>
