+++ SecBuzzerESM/Fluentd/conf/suricata.conf.bak
@@ -1,0 +1,97 @@
+<source>
+  @type tail
+  path /fluentd/suricata/eve*.json
+  pos_file /fluentd/Buffers/suricata.pos
+  read_from_head true
+  @label @SURICATA
+  tag suricata
+  <parse>
+    @type json
+  </parse>
+</source>
+<source>
+  @type tail
+  path /fluentd/dtm/eve-*.json
+  pos_file /fluentd/Buffers/dtm.pos
+  @label @SURICATA
+  tag suricata.file
+  <parse>
+    @type json
+  </parse>
+</source>
+<label @SURICATA>
+  <filter suricata.file>
+    @type record_transformer
+    auto_typecast true
+    enable_ruby true
+    <record>
+      log_time ${Time.strptime(record["timestamp"], "%Y-%m-%dT%H:%M:%S.%N%z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z')}
+      _log_type traffic
+      module DTM
+      dump_status 0
+      ticket 0
+    </record>
+  </filter>
+  <filter suricata>
+    @type record_transformer
+    auto_typecast true
+    enable_ruby true
+    <record>
+      log_time ${Time.strptime(record["timestamp"], "%Y-%m-%dT%H:%M:%S.%N%z").strftime('%Y-%m-%dT%H:%M:%S.%N%:z')}
+      _log_type traffic
+      module suricata
+      dump_status 0
+      ticket 0
+    </record>
+  </filter>
+  <match suricata.file>
+    @type elasticsearch
+    host localhost
+    port 19200
+    index_name dtm-%Y-%m
+    include_timestamp true
+    reload_on_failure true
+    reconnect_on_error true
+    with_transporter_log true
+    reload_connections false
+    request_timeout 15s
+    pipeline my_timestamp_pipeline
+    suppress_type_name true
+    <buffer time>
+      @type file
+      timekey 1d
+      flush_mode interval
+      flush_interval 30s
+      retry_max_interval 30s
+      path /fluentd/Buffers/ES_Buffers/dtm
+      chunk_limit_size 15M
+      total_limit_size 5G # default is 64G
+      timekey_use_utc false
+    </buffer>
+  </match>
+  <match suricata>
+    @type elasticsearch
+    host localhost
+    port 19200
+    index_name suricata-%Y-%m
+    include_timestamp true
+    reload_on_failure true
+    reconnect_on_error true
+    with_transporter_log true
+    reload_connections false
+    request_timeout 15s
+    pipeline my_timestamp_pipeline
+    suppress_type_name true
+    <buffer time>
+      @type file
+      timekey 1d
+      flush_mode interval
+      flush_interval 30s
+      retry_max_interval 30s
+      path /fluentd/Buffers/ES_Buffers/suricata
+      chunk_limit_size 15M
+      total_limit_size 5G # default is 64G
+      timekey_use_utc false
+    </buffer>
+  </match>
+</label>
