+++ SecBuzzerESM/Fluentd/docker-compose.yml
@@ -7,8 +7,12 @@ services:
             dockerfile: "Dockerfile"
         volumes:
             - "$FLUENTD_BUFFER_PATH:/fluentd/Buffers"
-            - "./conf/suricata.conf:/fluentd/etc/fluent.conf"
+            - "./conf/multiprocess.conf:/fluentd/etc/fluent.conf"
+            - "./conf/in_fluentd.conf:/fluentd/etc/in_fluentd.conf"
+            - "./conf/out_it_fluentd.conf:/fluentd/etc/out_it_fluentd.conf"
+            - "./conf/out_dtm_fluentd.conf:/fluentd/etc/out_dtm_fluentd.conf"
             - "/usr/share/zoneinfo/ROC:/etc/localtime:ro" 
             - "$SURICATA_LOG_PATH:/fluentd/suricata/"
+            - "$FLUENTD_LOG_PATH:/var/log/fluentd/"
         network_mode: host
         restart: unless-stopped
