+++ SecBuzzerESM/HISTORY.md
@@ -1,17 +1,4 @@
-# Release History
-# V2.3.4 (2021-06-25)
-- 再次更新 winlog模組的es query資料量上限
-
-# V2.3.3 (2021-06-24)
-- 更新 winlog模組的es query資料量上限
-
-# V2.3.2 (2021-06-21)
-- 更新文件: Winlogbeat設定SSL和CA.md
-
-# V2.3.1 (2021-06-15)
-- 移除compose.sh內的AI模組註解
-- 新增文件: SecBuzzerESM 連線安裝手冊.md
-
+# Release Histor
 # V2.3.0 (2021-06-05)
 - 修復 servicecheck.py 無法讀取環境變數
 - 修正 servicecheck 執行的時間
