+++ SecBuzzerESM/Install.sh
@@ -8,7 +8,6 @@ fi
 
 DOCKER_VERSION=18.09.3
 COMPOSE_VERSION=1.25.3
-GRAFANA_PIE_CHART_VERSION=1.5.0
 
 sudo apt update
 sudo apt install -y curl
@@ -27,11 +26,6 @@ get_script_dir () {
 }
 
 BASEDIR=$(get_script_dir)
-
-echo "[*] Download grafana plugin"
-rm -rf $BASEDIR/ES/grafana_plugins
-mkdir -p $BASEDIR/ES/grafana_plugins
-git -C $BASEDIR/ES/grafana_plugins/ clone https://github.com/grafana/piechart-panel.git --branch release-$GRAFANA_PIE_CHART_VERSION 
 
 if [ -f "/usr/bin/docker" ] && [ -f "/usr/local/bin/docker-compose" ]; then
     echo "[*] Already install docker!"
@@ -78,19 +72,19 @@ root hard nofile 655360
 
 find $BASEDIR -maxdepth 1 -mindepth 1 -type d -exec ln -s ../SecBuzzerESM.env {}/.env \; 2>/dev/null || true
 
-API_KEY=`cat $BASEDIR/SecBuzzerESM.env | grep API_KEY_VALUE | cut -d = -f 2`
-if [ -n "$API_KEY" ]
-then
-mkdir -p tmp/rules
-current_rules_version=`curl -X POST "https://test.api.secbuzzer.co/esmapi/web/file/fileVersion" -d "{'TypeCode': 'it'}" -H "Content-Type: application/json" -H "accept: */*" -H "authorization: $API_KEY" | cut -d : -f 2 | cut -d \" -f 2`
-curl -o rules.tgz "https://test.api.secbuzzer.co/esmapi/web/file/download/it/$current_rules_version" -H "accept: */*" -H "authorization: $API_KEY"
-tar zxvf rules.tgz -C tmp/rules
-sudo chown 1000:1000 tmp/* -R
-sudo rsync -r --delete tmp/rules/ Suricata/suricata/rules/
-rm -rf tmp rules.tgz
-else
-echo No API key found,Suricata rules download fail, check SecBuzzerESM.env
-fi
+#API_KEY=`cat $BASEDIR/SecBuzzerESM.env | grep API_KEY_VALUE | cut -d = -f 2`
+#if [ -n "$API_KEY" ]
+#then
+#mkdir -p tmp/rules
+#current_rules_version=`curl -X POST "https://api.esm.secbuzzer.co/esmapi/web/file/fileVersion" -d "{'TypeCode': 'it'}" -H "Content-Type: application/json" -H "accept: */*" -H "authorization: $API_KEY" | cut -d : -f 2 | cut -d \" -f 2`
+#curl -o rules.tgz "https://api.esm.secbuzzer.co/esmapi/web/file/download/it/$current_rules_version" -H "accept: */*" -H "authorization: $API_KEY"
+#tar zxvf rules.tgz -C tmp/rules
+#sudo chown 1000:1000 tmp/* -R
+#sudo rsync -r --delete tmp/rules/ Suricata/suricata/rules/
+#rm -rf tmp rules.tgz
+#else
+#echo No API key found,Suricata rules download fail, check SecBuzzerESM.env
+#fi
 
 mkdir -p /opt/Logs/ES/volume/es
 mkdir -p /opt/Logs/Suricata
@@ -105,6 +99,8 @@ sudo docker network create esm_network 2>/dev/null || true
 
 echo "Disable Swap"
 swapoff -a
+rm -rf /swap.img
 sed -i 's/.*swap.*/#&/' /etc/fstab
+
 
 echo "Done!"
