+++ SecBuzzerESM/SecBuzzerESM.env
@@ -1,6 +1,6 @@
 # === Elasticsearch === 
 ES_VOLUME_PATH=/opt/Logs/ES/volume
-ES_MEM=4096m
+ES_MEM=8192m
 # === Fluentd ===
 FLUENTD_LOG_PATH=/opt/Logs/Fluentd/Logs
 FLUENTD_BUFFER_PATH=/opt/Logs/Buffers
@@ -23,5 +23,3 @@ FLOWSCAN_SOURCEIP=
 FLOWSCAN_DESTIP=
 FLOWSCAN_TOTALCOUNT=1000
 NMAPSCAN_IP=
-# === DEV ===
-DEV_MODE=True
