--- SecBuzzerESM/Suricata/suricata/rules/.rules_folder.md
@@ -1,1 +1,0 @@
-:ghost:
