+++ SecBuzzerESM/compose.sh
@@ -11,7 +11,7 @@ docker-compose --env-file SecBuzzerESM.env -f ES/docker-compose.yml $cmd
 docker-compose --env-file SecBuzzerESM.env -f Crontab/docker-compose.yml $cmd
 docker-compose --env-file SecBuzzerESM.env -f Fluentd/docker-compose.yml $cmd
 docker-compose --env-file SecBuzzerESM.env -f Suricata/docker-compose.yml $cmd
-docker-compose --env-file SecBuzzerESM.env -f AI/docker-compose.yml $cmd
+# docker-compose --env-file SecBuzzerESM.env -f AI/docker-compose.yml $cmd
 docker-compose --env-file SecBuzzerESM.env -f WEB/docker-compose.yml $cmd
-docker-compose --env-file SecBuzzerESM.env -f Packetbeat/docker-compose.yml $cmd
-docker-compose --env-file SecBuzzerESM.env -f Scan/docker-compose.yml $cmd
+# docker-compose --env-file SecBuzzerESM.env -f Packetbeat/docker-compose.yml $cmd
+# docker-compose --env-file SecBuzzerESM.env -f Scan/docker-compose.yml $cmd
